fixing issues when updating and deleting notes

+++ src/modules/note/note.controller.js
@@ -1,19 +1,31 @@
 import { noteModel } from "../../../Database/models/noteModel.js";
 import mongoose from 'mongoose';
 const addNote = async (req, res) => {
-    if (req.body.createdBy.toString() !== req.userId) {
-        return res.json({ message: "error happened while adding tthe note try again later!!" });
-    }
+    req.body.createdBy = req.userId;
     await noteModel.insertMany(req.body);
     res.json({ message: "note was added successfully" });
 };
 
 const updateNote = async (req, res) => {
-    note = await noteModel.findByIdAndUpdate(req.params.id, req.body, { new: true });
+    let noteExist = await noteModel.findById(req.params.id);
+    if (!noteExist) {
+        return res.status(404).json({ message: "note not found" });
+    }
+    if (req.userId !== noteExist.createdBy._id.toString()) {
+        return res.status(403).json({ message: "forbidden" });
+    }
+    let note = await noteModel.findByIdAndUpdate(req.params.id, req.body, { new: true });
     res.json({ message: "Success", note });
 };
 
 const deleteNote = async (req, res) => {
+    let noteExist = await noteModel.findById(req.params.id);
+    if (!noteExist) {
+        return res.status(404).json({ message: "note not found" });
+    }
+    if (req.userId !== noteExist.createdBy._id.toString()) {
+        return res.status(403).json({ message: "forbidden" });
+    }
     await noteModel.findByIdAndDelete(req.params.id);
     res.json({ message: "Success" });
 };
